Corregido errores y hecha primera version de astar (no cuenta transbordo)

+++ lib/backend/edge.dart
@@ -4,7 +4,7 @@ class Edge {
 
     final Vertex _station1;
     final Vertex _station2;
-    final int _time;    //time que se tarda en recorrerlo
+    final double _time;    //time que se tarda en recorrerlo
 
     Edge (this._station1, this._station2, this._time);
 
@@ -19,5 +19,5 @@ class Edge {
         throw Exception("Error: Se ha pedido la siguiente estacion de ($estacion) un nodo que no se encuentra en la arista");
     }
 
-    int gettime () => _time;
+    double gettime () => _time;
 }
